basic kernel running with user program
