+++ project2/os.c
@@ -1,6 +1,228 @@
 #include "os.h"
 
+#define LOW_BYTE(X) (((uint16_t)X) & 0xFF)
+#define HIGH_BYTE(X) ((((uint16_t)X) >> 8) & 0xFF)
+#define ZeroMemory(X, N) memset(&X, 0, N)
+
+/**
+ *  This is the set of states that a task can be in at any given time.
+ */
+typedef enum process_states
+    {
+     DEAD = 0,
+     READY,
+     RUNNING
+    } PROCESS_STATES;
+
+/**
+ * This is the set of kernel requests, i.e., a request code for each system call.
+ */
+typedef enum kernel_request_type
+    {
+     NONE = 0,
+     CREATE,
+     NEXT,
+     TERMINATE
+    } KERNEL_REQUEST_TYPE;
+
+/**
+ * Each task is represented by a process descriptor, which contains all
+ * relevant information about this task. For convenience, we also store
+ * the task's stack, i.e., its workspace, in here.
+ */
+typedef struct ProcessDescriptor
+    {
+    unsigned char *sp;   /* stack pointer into the "workSpace" */
+    unsigned char workSpace[WORKSPACE];
+    PROCESS_STATES state;
+    voidfuncptr  code;   /* function to be executed as a task */
+    KERNEL_REQUEST_TYPE request;
+    } PD;
+
+/**
+ * This internal kernel function is the context switching mechanism.
+ * It is done in a "funny" way in that it consists two halves: the top half
+ * is called "Exit_Kernel()", and the bottom half is called "Enter_Kernel()".
+ * When kernel calls this function, it starts the top half (i.e., exit). Right in
+ * the middle, "Cp" is activated; as a result, Cp is running and the kernel is
+ * suspended in the middle of this function. When Cp makes a system call,
+ * it enters the kernel via the Enter_Kernel() software interrupt into
+ * the middle of this function, where the kernel was suspended.
+ * After executing the bottom half, the context of Cp is saved and the context
+ * of the kernel is restore. Hence, when this function returns, kernel is active
+ * again, but Cp is not running any more.
+ * (See file "cswitch.S" for details.)
+ */
+extern void CSwitch();
+extern void Exit_Kernel();    /* this is the same as CSwitch() */
+
+void Task_Terminate(void);
+
+/**
+ * This external function could be implemented in two ways:
+ *  1) as an external function call, which is called by Kernel API call stubs;
+ *  2) as an inline macro which maps the call into a "software interrupt";
+ *       as for the AVR processor, we could use the external interrupt feature,
+ *       i.e., INT0 pin.
+ *  Note: Interrupts are assumed to be disabled upon calling Enter_Kernel().
+ *     This is the case if it is implemented by software interrupt. However,
+ *     as an external function call, it must be done explicitly. When Enter_Kernel()
+ *     returns, then interrupts will be re-enabled by Enter_Kernel().
+ */
+extern void Enter_Kernel();
+
+/**
+ * This table contains ALL process descriptors. It doesn't matter what
+ * state a task is in.
+ */
+// TODO: Replace with Priority based queues
+static PD Process[MAXTHREAD];
+
+/**
+ * The process descriptor of the currently RUNNING task.
+ */
+volatile static PD* Cp;
+
+/**
+ * Since this is a "full-served" model, the kernel is executing using its own
+ * stack. We can allocate a new workspace for this kernel stack, or we can
+ * use the stack of the "main()" function, i.e., the initial C runtime stack.
+ * (Note: This and the following stack pointers are used primarily by the
+ *   context switching code, i.e., CSwitch(), which is written in assembly
+ *   language.)
+ */
+volatile unsigned char *KernelSp;
+
+/**
+ * This is a "shadow" copy of the stack pointer of "Cp", the currently
+ * running task. During context switching, we need to save and restore
+ * it into the appropriate process descriptor.
+ */
+volatile unsigned char *CurrentSp;
+
+/** index to next task to run */
+volatile static unsigned int NextP;
+
+/** 1 if kernel has been started; 0 otherwise. */
+volatile static unsigned int KernelActive;
+
+/** number of tasks created so far */
+volatile static unsigned int Tasks;
+
 void OS_Abort(unsigned int error) {}
+
+void Kernel_Task_Create_At(PD *p, voidfuncptr f) {
+    uint8_t *sp = &(p->workSpace[WORKSPACE - 1]);
+
+    //Clear the contents of the workspace
+    ZeroMemory(p->workSpace, WORKSPACE);
+
+    //Notice that we are placing the address (16-bit) of the functions
+    //onto the stack in reverse byte order (least significant first, followed
+    //by most significant).  This is because the "return" assembly instructions
+    //(rtn and rti) pop addresses off in BIG ENDIAN (most sig. first, least sig.
+    //second), even though the AT90 is LITTLE ENDIAN machine.
+
+    //Store terminate at the bottom of stack to protect against stack underrun.
+    *sp-- = LOW_BYTE(Task_Terminate);
+    *sp-- = HIGH_BYTE(Task_Terminate);
+    *sp-- = LOW_BYTE(0);
+
+    //Place return address of function at bottom of stack
+    *sp-- = LOW_BYTE(f);
+    *sp-- = HIGH_BYTE(f);
+    *sp-- = LOW_BYTE(0);
+//Place stack pointer at top of stack
+    sp = sp - 34;
+
+    p->sp = sp;      /* stack pointer into the "workSpace" */
+    p->code = f;     /* function to be executed as a task */
+    p->request = NONE;
+    p->state = READY;
+}
+
+/**
+ * This internal kernel function is a part of the "scheduler". It chooses the
+ * next task to run, i.e., Cp.
+ */
+static void Dispatch() {
+    /* find the next READY task
+     * Note: if there is no READY task, then this will loop forever!.
+     */
+    while(Process[NextP].state != READY) {
+        NextP = (NextP + 1) % MAXTHREAD;
+    }
+
+    Cp = &(Process[NextP]);
+    CurrentSp = Cp->sp;
+    Cp->state = RUNNING;
+
+    NextP = (NextP + 1) % MAXTHREAD;
+}
+
+static void Next_Kernel_Request() {
+    Dispatch();  /* select a new task to run */
+
+    while(1) {
+        Cp->request = NONE; /* clear its request */
+
+        /* activate this newly selected task */
+        CurrentSp = Cp->sp;
+        Exit_Kernel();    /* or CSwitch() */
+
+        /* if this task makes a system call, it will return to here! */
+
+        /* save the Cp's stack pointer */
+        Cp->sp = CurrentSp;
+
+        switch(Cp->request){
+        case CREATE:
+            Task_Create( Cp->code );
+            break;
+        case NEXT:
+        case NONE:
+            /* NONE could be caused by a timer interrupt */
+            Cp->state = READY;
+            Dispatch();
+            break;
+        case TERMINATE:
+            /* deallocate all resources used by this task */
+            Cp->state = DEAD;
+            Dispatch();
+            break;
+        default:
+            /* Houston! we have a problem here! */
+            break;
+        }
+    }
+}
+
+/**
+ * The calling task terminates itself.
+ */
+void Task_Terminate()
+{
+    if (KernelActive) {
+        OS_DI();
+        Cp->request = TERMINATE;
+        Enter_Kernel();
+        /* never returns here! */
+    }
+}
+
+void Task_Create(voidfuncptr f) {
+    int x;
+
+    if (Tasks == MAXTHREAD) return;  /* Too many task! */
+
+    /* find a DEAD PD that we can use  */
+    for (x = 0; x < MAXTHREAD; x++) {
+        if (Process[x].state == DEAD) break;
+    }
+
+    ++Tasks;
+    Kernel_Task_Create_At( &(Process[x]), f );
+}
 
 PID   Task_Create_System(void (*f)(void), int arg) {}
 
@@ -8,7 +230,13 @@ PID   Task_Create_RR(    void (*f)(void), int arg) {}
 
 PID   Task_Create_Period(void (*f)(void), int arg, TICK period, TICK wcet, TICK offset) {}
 
-void Task_Next(void) {}
+void Task_Next(void) {
+    if (KernelActive) {
+        OS_DI();
+        Cp->request = NEXT;
+        Enter_Kernel();
+    }
+}
 
 int  Task_GetArg(void) {}
 
@@ -23,3 +251,29 @@ void Msg_Rply( PID  id,          unsigned int r ) {}
 void Msg_ASend( PID  id, MTYPE t, unsigned int v ) {}
 
 unsigned int Now() {}
+
+void OS_Init() {
+    int x;
+
+    Tasks = 0;
+    KernelActive = 0;
+    NextP = 0;
+
+    //Reminder: Clear the memory for the task on creation.
+    for (x = 0; x < MAXTHREAD; x++) {
+        memset(&(Process[x]),0,sizeof(PD));
+        Process[x].state = DEAD;
+    }
+}
+
+void OS_Start() {
+    if ( (! KernelActive) && (Tasks > 0)) {
+        OS_DI();
+        /* we may have to initialize the interrupt vector for Enter_Kernel() here. */
+
+        /* here we go...  */
+        KernelActive = 1;
+        Next_Kernel_Request();
+        /* NEVER RETURNS!!! */
+    }
+}
